requirements.txt
edge_tts==7.0.0
moviepy==1.0.3
openai==1.61.0
openai_whisper==20240930
pandas==2.2.3
Pillow==11.1.0
praw==7.8.1
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.28.1
